pyproject.toml
[project]
name = "squadron"
version = "0.1.0"
description = "GitHub-Native multi-LLM-agent autonomous development framework"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "httpx>=0.27",
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "aiosqlite>=0.20",
    "github-copilot-sdk>=0.1",
    "PyJWT[crypto]>=2.0",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "respx>=0.22",
    "ruff>=0.5",
]

[project.scripts]
squadron = "squadron.__main__:main"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "live: tests that hit real LLM APIs via live inference (30-120s each)",
]
